Sync the FreeBSD port with upstream changes.
The buffer length in sendCmd() has changed from a uint16_t to a uint32_t.

diskScan() is now implemented in each port.  The FreeBSD diskScan is
just a shell for now, I will implement a real scan later.

+++ freebsd/DtaDevFreeBSDNvme.cpp
@@ -49,7 +49,7 @@ bool DtaDevFreeBSDNvme::init(const char * devref)
 
 /** Send an ioctl to the device using nvme admin commands. */
 uint8_t DtaDevFreeBSDNvme::sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
-                         void * buffer, uint16_t bufferlen)
+                         void * buffer, uint32_t bufferlen)
 {
 	struct nvme_pt_command	pt;
 	int err;
