+++ freebsd/DtaDevFreeBSDSata.h
@@ -49,7 +49,7 @@ public:
      * @param bufferlen length of the input/output buffer
      */
     uint8_t sendCmd(ATACOMMAND cmd, uint8_t protocol, uint16_t comID,
-            void * buffer, uint16_t bufferlen);
+            void * buffer, uint32_t bufferlen);
     void identify(OPAL_DiskInfo& disk_info);
     struct cam_device *camdev;
     int isSAS; /* The device is sas */
